+++ src/main.jsx
@@ -4,10 +4,10 @@ import App from "./App.jsx";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import { Contact } from "./page/Contact/Contact.jsx";
 import "./index.css";
-import "../firebase.js"
-import { ContactAdmin, } from "./page/admin/contacto-admin/ContactAdmin.jsx";
+import "../firebase.js";
+import { ContactAdmin } from "./page/admin/contacto-admin/ContactAdmin.jsx";
 import { AdminManagement } from "./page/admin/managementUser/adminManagement.jsx";
-import { CourseCard } from "./component/CourseCard.jsx";
+import { CourseCard } from "./component/CourseCard/CourseCard.jsx";
 import { AdminNavBar } from "./component/AdminNavBar.jsx";
 
 const router = createBrowserRouter([
@@ -34,8 +34,7 @@ const router = createBrowserRouter([
   {
     path: "adminNavBar",
     element: <AdminNavBar />,
-  }
-
+  },
 ]);
 
 ReactDOM.createRoot(document.getElementById("root")).render(
